html/semantics/invokers/invokeevent-dispatch-shadow.tentative.html

<!doctype html>
<meta charset="utf-8" />
<meta name="author" title="Keith Cirkel" href="mailto:keithamus@github.com" />
<link rel="help" href="https://open-ui.org/components/invokers.explainer/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/invoker-utils.js"></script>

<div id="div"></div>
<button id="button"></button>

<script>
  test(function () {
    const host = document.createElement("div");
    const child = host.appendChild(document.createElement("p"));
    const shadow = host.attachShadow({ mode: "closed" });
    const slot = shadow.appendChild(document.createElement("slot"));
    let childEvent = null;
    let childEventTarget = null;
    let childEventRelatedTarget = null;
    let hostEvent = null;
    slot.addEventListener(
      "invoke",
      (e) => {
        childEvent = e;
        childEventTarget = e.target;
        childEventRelatedTarget = e.relatedTarget;
      },
      { once: true },
    );
    host.addEventListener(
      "invoke",
      (e) => {
        hostEvent = e;
      },
      { once: true },
    );
    const event = new InvokeEvent("invoke", {
      bubbles: true,
      relatedTarget: slot,
      composed: true,
    });
    slot.dispatchEvent(event);
    assert_true(childEvent instanceof InvokeEvent, "slot saw invoke event");
    assert_equals(
      childEventTarget,
      slot,
      "target is child inside shadow boundary",
    );
    assert_equals(
      childEventRelatedTarget,
      slot,
      "relatedTarget is child inside shadow boundary",
    );
    assert_equals(
      hostEvent,
      null,
      "event dispatch not fired outside shadow-tree",
    );
  }, "InvokeEvent does not bubble across shadow boundaries because relatedTarget is set");

  test(function () {
    const host = document.createElement("div");
    const shadow = host.attachShadow({ mode: "closed" });
    const button = shadow.appendChild(document.createElement("button"));
    const invokee = host.appendChild(document.createElement("div"));
    button.invokeTargetElement = invokee;
    let event = null;
    let eventTarget = null;
    let eventRelatedTarget = null;
    invokee.addEventListener(
      "invoke",
      (e) => {
        event = e;
        eventTarget = e.target;
        eventRelatedTarget = e.relatedTarget;
      },
      { once: true },
    );
    clickOn(button);
    assert_true(event instanceof InvokeEvent);
    assert_equals(eventTarget, invokee, "target is invokee");
    assert_equals(eventRelatedTarget, host, "relatedTarget is host");
  }, "cross shadow InvokeEvent retargets relatedTarget to host element");
</script>
